+++ gen-models/src/annotations.rs
@@ -1,5 +1,4 @@
 use std::{
-    cmp::{max, min},
     collections::HashMap,
     path::{Path, PathBuf},
     rc::Rc,
@@ -7,10 +6,8 @@ use std::{
 
 use anyhow::anyhow;
 use gen_core::{
-    HashId, NodeIntervalBlock, Strand, calculate_hash,
+    HashId, NodeIntervalBlock, calculate_hash,
     config::Workspace,
-    path::PathBlock,
-    range::{OrderedMerge, Range, merge_ordered_items},
     region::{Region, RegionResolutionError, RegionResolver},
     traits::Capnp,
 };
@@ -172,130 +169,6 @@ pub struct BedExtra {
     pub block_sizes: Option<Vec<i64>>,
     pub block_starts: Option<Vec<i64>>,
     pub other_fields: Vec<String>,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct AnnotationSegment {
-    pub node_id: HashId,
-    pub range: Range,
-    pub strand: Strand,
-}
-
-impl OrderedMerge for AnnotationSegment {
-    fn should_merge_with(&self, next: &Self) -> bool {
-        self.node_id == next.node_id
-            && self.strand == next.strand
-            && next.range.start >= self.range.start
-            && next.range.start <= self.range.end
-    }
-
-    fn merge_with(&mut self, next: &Self) {
-        self.range.end = max(self.range.end, next.range.end);
-    }
-}
-
-pub fn accession_edges_to_segments(
-    edges: &[crate::accession::AccessionEdge],
-) -> Vec<AnnotationSegment> {
-    let mut segments = Vec::new();
-    let mut current_node = None;
-    let mut current_start = None;
-    let mut current_strand = None;
-
-    for edge in edges {
-        if edge.source_coordinate < 0 {
-            current_node = Some(edge.target_node_id);
-            current_start = Some(edge.target_coordinate);
-            current_strand = Some(edge.target_strand);
-            continue;
-        }
-
-        if let (Some(node_id), Some(start), Some(strand)) =
-            (current_node, current_start, current_strand)
-        {
-            let (segment_start, segment_end) = if start <= edge.source_coordinate {
-                (start, edge.source_coordinate)
-            } else {
-                (edge.source_coordinate, start)
-            };
-            if segment_end > segment_start {
-                segments.push(AnnotationSegment {
-                    node_id,
-                    range: Range {
-                        start: segment_start,
-                        end: segment_end,
-                    },
-                    strand,
-                });
-            }
-        }
-
-        if edge.target_coordinate < 0 {
-            break;
-        }
-
-        current_node = Some(edge.target_node_id);
-        current_start = Some(edge.target_coordinate);
-        current_strand = Some(edge.target_strand);
-    }
-
-    segments
-}
-
-pub fn project_annotation_segments(
-    accession_segments: &[AnnotationSegment],
-    path_blocks: &[PathBlock],
-    preserve_part_boundaries: bool,
-) -> Vec<AnnotationSegment> {
-    let projected = accession_segments
-        .iter()
-        .flat_map(|segment| {
-            merge_ordered_items(
-                path_blocks
-                    .iter()
-                    .filter_map(|block| {
-                        if block.node_id != segment.node_id {
-                            return None;
-                        }
-                        let overlap_start = max(segment.range.start, block.sequence_start);
-                        let overlap_end = min(segment.range.end, block.sequence_end);
-                        if overlap_end <= overlap_start {
-                            return None;
-                        }
-
-                        let (start, end) = if block.strand == Strand::Reverse {
-                            (
-                                block.path_start + (block.sequence_end - overlap_end),
-                                block.path_start + (block.sequence_end - overlap_start),
-                            )
-                        } else {
-                            (
-                                block.path_start + (overlap_start - block.sequence_start),
-                                block.path_start + (overlap_end - block.sequence_start),
-                            )
-                        };
-                        let strand = if block.strand == Strand::Reverse {
-                            segment.strand.complement()
-                        } else {
-                            segment.strand
-                        };
-
-                        Some(AnnotationSegment {
-                            node_id: block.node_id,
-                            range: Range { start, end },
-                            strand,
-                        })
-                    })
-                    .collect(),
-            )
-        })
-        .collect::<Vec<_>>();
-
-    if preserve_part_boundaries {
-        projected
-    } else {
-        merge_ordered_items(projected)
-    }
 }
 
 fn serialize_annotation_extra(extra: Option<&AnnotationExtra>) -> Result<String, AnnotationError> {
@@ -1038,7 +911,7 @@ impl AnnotationFile {
 mod tests {
     use std::fs;
 
-    use gen_core::{HashId, Strand, path::PathBlock, range::Range, region::RegionResolutionError};
+    use gen_core::{HashId, region::RegionResolutionError};
 
     use super::*;
     use crate::{
@@ -1193,66 +1066,6 @@ mod tests {
         );
         assert_eq!(reparsed.gff, None);
         assert_eq!(reparsed.bed, None);
-    }
-
-    #[test]
-    fn merges_overlapping_annotation_segments() {
-        let node_id = HashId::convert_str("node");
-        let merged = merge_ordered_items(vec![
-            AnnotationSegment {
-                node_id,
-                range: Range { start: 2, end: 5 },
-                strand: Strand::Forward,
-            },
-            AnnotationSegment {
-                node_id,
-                range: Range { start: 4, end: 8 },
-                strand: Strand::Forward,
-            },
-        ]);
-
-        assert_eq!(
-            merged,
-            vec![AnnotationSegment {
-                node_id,
-                range: Range { start: 2, end: 8 },
-                strand: Strand::Forward,
-            }]
-        );
-    }
-
-    #[test]
-    fn project_annotation_segments_complements_reverse_blocks() {
-        let node_id = HashId::convert_str("node");
-        let projected = project_annotation_segments(
-            &[AnnotationSegment {
-                node_id,
-                range: Range { start: 10, end: 20 },
-                strand: Strand::Forward,
-            }],
-            &[PathBlock {
-                node_id: HashId::convert_str("node"),
-                block_sequence: String::new(),
-                sequence_start: 0,
-                sequence_end: 30,
-                path_start: 100,
-                path_end: 130,
-                strand: Strand::Reverse,
-            }],
-            false,
-        );
-
-        assert_eq!(
-            projected,
-            vec![AnnotationSegment {
-                node_id,
-                range: Range {
-                    start: 110,
-                    end: 120,
-                },
-                strand: Strand::Reverse,
-            }]
-        );
     }
 
     #[test]
